requirements.txt
asgiref==3.5.0
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.12
click==8.0.4
coreapi==2.3.3
coreschema==0.0.4
cryptography==36.0.2
defusedxml==0.7.1
dj-rest-auth==2.2.3
Django==4.0.3
django-allauth==0.49.0
djangorestframework==3.13.1
djangorestframework-jwt==1.11.0
djangorestframework-simplejwt==5.1.0
drf-yasg==1.20.0
flex==6.14.1
idna==3.3
inflection==0.5.1
itypes==1.2.0
Jinja2==3.0.3
jsonpointer==2.2
MarkupSafe==2.1.0
oauthlib==3.2.0
packaging==21.3
pycparser==2.21
PyJWT==1.7.1
pyparsing==3.0.7
python3-openid==3.2.0
pytz==2021.3
PyYAML==6.0
requests==2.27.1
requests-oauthlib==1.3.1
rfc3987==1.3.8
ruamel.yaml==0.17.21
ruamel.yaml.clib==0.2.6
six==1.16.0
sqlparse==0.4.2
strict-rfc3339==0.7
uritemplate==4.1.1
urllib3==1.26.8
validate-email==1.3
